revno: 3114 message:   Fix bug: Calculated-data-element-decimal-factors-are-truncated-in-display.
committer: Tran Chau <tran.hispvietnam@gmail.com>
branch nick: dhis2
timestamp: Tue 2011-03-22 12:49:48 +0700

+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataelement/DataElementService.java
@@ -340,7 +340,7 @@ public interface DataElementService
      * @return a map of factors for the DataElements in the given
      *         CalculatedDataElement
      */
-    Map<DataElement, Integer> getDataElementFactors( CalculatedDataElement calculatedDataElement );
+    Map<DataElement, Double> getDataElementFactors( CalculatedDataElement calculatedDataElement );
 
     /**
      * Returns a Map of factors for the Operands in the given
@@ -350,7 +350,7 @@ public interface DataElementService
      * @return a map of factors for the Operands in the given
      *         CalculatedDataElement
      */
-    Map<String, Integer> getOperandFactors( CalculatedDataElement calculatedDataElement );
+    Map<String, Double> getOperandFactors( CalculatedDataElement calculatedDataElement );
 
     /**
      * Returns a collection of OperandIds in the given CalculatedDataElement
